+++ crest/Assets/Crest/Crest/Scripts/BuildCommandBuffer.cs
@@ -111,6 +111,12 @@ namespace Crest
 
             BuildLodData(OceanRenderer.Instance, _buf);
 
+            if (OceanRenderer.Instance.ViewCamera != null)
+            {
+                // Fixes flickering non mesh renderer renderers (like particles). Method is undocumented.
+                Camera.SetupCurrent(OceanRenderer.Instance.ViewCamera);
+            }
+
             // This will execute at the beginning of the frame before the graphics queue
             Graphics.ExecuteCommandBuffer(_buf);